Encoding IDs in GetTaskById

+++ src/Core/Nexus.DEB.Application/Common/Models/Task/TaskDetail.cs
@@ -3,7 +3,7 @@
 	public class TaskDetail : EntityDetailBase
 	{
 		public DateTime? DueDate { get; set; }
-		public short TaskTypeId { get; set; }
+		public int TaskTypeId { get; set; }
 		public string TaskType { get; set; } = string.Empty;
 		public int? ActivityId { get; set; }
 		public string Status { get; set; }
